update Rprofile to display API time; clean pkgs
Since the cache may occasionally go out of date because of reasons,
displaying the time it was last updated helps me know if I need to ping
Scott about it.

I've also figured out the right order in which I'm supposed to unload
the packages that I imported from the three I use.